Consolidated shared elements of python topic tool nodes
Both existing python nodes as well as the planned filter relay node use
these shared elements. Using a shared parent class cleans up code
repetition.
